# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: fashion
    description: Fashion item operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /fashion/items:
    get:
      operationId: getFashionItems
      tags: [fashion]
      summary: Get men's fashion items
      description: Fetches men's fashion items from H&M, optionally filtered by budget
      parameters:
        - name: budget
          in: query
          required: false
          schema:
            type: number
          description: Max price filter
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of fashion items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FashionItemsResponse"
        "502":
          description: Upstream API error
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    FashionItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        formattedPrice:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        buyUrl:
          type: string
      required:
        - id
        - name
        - price
        - formattedPrice
        - imageUrl
        - category
        - buyUrl
    FashionItemsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FashionItem"
      required:
        - items