# lib/api-zod/src/generated/types/fashionItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FashionItem {
  id: string;
  name: string;
  price: number;
  formattedPrice: string;
  imageUrl: string;
  category: string;
  buyUrl: string;
}